feat: refactor tax ID validation logic for Spain and Portugal
